* commit '28521177fa9c04593a3e2c017a780a1076231dfe':
  Connect of backend to graph drawing

+++ EnergyPrediction/EnergyPrediction/Genetic Program/TreeUtil/VisitorPattern.cs
@@ -90,7 +90,7 @@ namespace EnergyPrediction
             return aXThreshold;
         }
 
-        public static double Calculate(TreeNode<MathObject> root, int x)
+        public static double Calculate(TreeNode<MathObject> root, double x)
         {
             if (root.Data.GetType().Equals(typeof(MathNumber)))
             {
